Moved relationship components to a relationship folder.
